Обновление страницы public
- Добавление индикатора числа лайков на карточке
- Изменение положения кнопки "Добавить карточку"
