+++ dogs_app/frontend_dogs/src/component/UI/Card/Card.jsx
@@ -11,9 +11,14 @@ const Card = ({img_path, title, likes}) => {
             </div>
             <div className="row g-0">
                 <div className="col-auto">
-                    <div className="card-body" style={{"z-index": '3'}}>
+                    <div className="card-body" style={{"zIndex": '3'}}>
                         <h5 className="card-title">{title}</h5>
-                        <p> {likes} </p>
+                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="Pink" viewBox="0 0 20 20"
+                             className="bi bi-suit-heart-fill m-3" style={{position: "absolute"}} >
+                            <path
+                                d="M4 1c2.21 0 4 1.755 4 3.92C8 2.755 9.79 1 12 1s4 1.755 4 3.92c0 3.263-3.234 4.414-7.608 9.608a.513.513 0 0 1-.784 0C3.234 9.334 0 8.183 0 4.92 0 2.755 1.79 1 4 1z"/>
+                        </svg>
+                        <p style={{color: "gray"}}> {likes} </p>
                     </div>
                 </div>
             </div>
